[8.4] [MOD-18138] Serve queries whose union or intersection has more than 65535 children (#11256)
* [MOD-18138] Serve queries whose union or intersection has more than 65535 children (#11238)

(cherry picked from commit 2aba0b2c35261186253eeff9d844a1c5739bbc9d)

Adapted for 8.4: this branch predates the standalone `index_result` crate, so the
capacity widening lands in its own layout, and the iterator-level tests are
replaced by an aggregate-level one where the Rust union iterator does not exist.


* Update the mem_usage doctest for the wider header


---------

+++ src/redisearch_rs/headers/low_memory_thin_vec.h
@@ -5,7 +5,7 @@
 /**
  * The type used to represent the capacity of a `LowMemoryThinVec`.
  */
-typedef uint16_t SizeType;
+typedef uint32_t SizeType;
 
 /**
  * The header of a LowMemoryThinVec.
